model/requirements.txt
torch==2.0.1
torchvision==0.15.2
Pillow==9.3.0
Flask==2.3.3
Flask-Cors==4.0.0
numpy==1.24.1
qiskit==0.44.1
qiskit-aer==0.12.2
keras==2.13.1
tensorflow==2.13.0
gunicorn